pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "pybind11"]
build-backend = "setuptools.build_meta"

[project]
name = "ap"
version = "0.0.1"
description = "Anthropomorphic Phantoms"
authors = [
  { name="Kris Dreher", email="k.dreher@dkfz-heidelberg.de" },
]
readme = "README.md"
requires-python = ">=3.8,<3.12"


# Requirements
dependencies = [
    "matplotlib",
    "numpy",
    "scipy",
    "pandas",
    "simpa",
    "pynrrd",
    "patato@git+https://github.com/BohndiekLab/patato",
]

[tool.setuptools.packages.find]
# All the following settings are optional:
include = ["*"]  # ["*"] by default

[project.scripts]
